+++ Automate/Framework/Machines/Objects/LightningRodMachine.cs
@@ -1,17 +1,19 @@
+using StardewValley;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
 {
     /// <summary>A lightning rod that accepts input and provides output.</summary>
-    internal class LightningRodMachine : GenericMachine
+    internal class LightningRodMachine : GenericObjectMachine<SObject>
     {
         /*********
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="machine">The underlying machine.</param>
-        public LightningRodMachine(SObject machine)
-            : base(machine) { }
+        /// <param name="location">The location containing the machine.</param>
+        public LightningRodMachine(SObject machine, GameLocation location)
+            : base(machine, location) { }
 
         /// <summary>Get the output item.</summary>
         public override ITrackedStack GetOutput()