+++ Automate/Framework/Machines/Objects/SeedMakerMachine.cs
@@ -6,7 +6,7 @@ using SObject = StardewValley.Object;
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
 {
     /// <summary>A seed maker that accepts input and provides output.</summary>
-    internal class SeedMakerMachine : GenericMachine
+    internal class SeedMakerMachine : GenericObjectMachine<SObject>
     {
         /*********
         ** Properties
@@ -20,8 +20,9 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="machine">The underlying machine.</param>
-        public SeedMakerMachine(SObject machine)
-            : base(machine) { }
+        /// <param name="location">The location containing the machine.</param>
+        public SeedMakerMachine(SObject machine, GameLocation location)
+            : base(machine, location) { }
 
         /// <summary>Provide input to the machine.</summary>
         /// <param name="input">The available items.</param>